Prepare checkout

+++ catalog_engine/catalog_cart.py
@@ -49,7 +49,7 @@ class CatalogCart():
             raise Exception('Invalid entry key size')
         return entry_key, index
 
-    def get_cart_items(self, cart_id, limit=100, backend=False):
+    def get_cart_items(self, cart_id, limit=100):
         ct = nsql.table('client_cart_item').get(
             select = 'count(id) as ct',
             where = {'cart_id': cart_id},
@@ -67,10 +67,11 @@ class CatalogCart():
                 'label',
                 'price',
                 'quantity',
+                'net_tax',
+                'backend_id',
                 'option_data',
                 'image_url as image',
                 '(price * quantity) as total',
-                'backend_id',
                 '(select user_backend.user_id from user_backend where user_backend.id=backend_id) as user_id',
             ],
             where = {'cart_id': cart_id},
@@ -80,13 +81,10 @@ class CatalogCart():
         items = []
         mrch_key = {}
         for r in q:
-            if backend:
-                r['entry_key'] = r['id']
             encoder = krock32.Encoder(checksum=False)
             encoder.update(r['id'])
             r['id'] = encoder.finalize().upper() + '.' + str(r['product_index'])
-            if not backend:
-                del r['product_index']
+            del r['product_index']
             r['option_data'] = json.loads(r['option_data'])
             items.append(r)
             mrch_key.setdefault(r['user_id'], [])
@@ -147,6 +145,9 @@ class CatalogCart():
         tax_diff = None
         if backend == 'vendure':
             vb = VendureBackend(None, URIRef(urec['merchant_uri']), bkcfg['vendure_url'])
+            current_data = internal_data['backend'][backend][backend_id]
+            if 'auth' in current_data:
+                vb.set_auth_token(current_data['auth'])
             if product['type'] == 'IProductGroup':
                 variant_id = product['hasVariant'][index]['productID']
             else:
@@ -173,14 +174,55 @@ class CatalogCart():
         cart.update(updates)
         return item_data, tax_diff
 
-    def backend_remove_cart_item(self, cart, backend_id, product, index, price, quantity):
+    def backend_update_cart_item(self, cart, backend_id, cart_item, quantity):
         bkrec = sql_row('user_backend', id=backend_id)
         if not bkrec.exists():
             raise Exception('Invalid user backend: {}'.format(bkid))
         urec = sql_row('user', id=bkrec['user_id'])
         bkcfg = json.loads(bkrec['config_data'])
         backend = bkrec['backend_name']
+        internal_data = json.loads(cart['cart_data'])
+        net_price = None
+        tax_diff = None
+        if backend == 'vendure':
+            backend_data = internal_data['backend'][backend][backend_id]
+            vb = VendureBackend(None, URIRef(urec['merchant_uri']), bkcfg['vendure_url'])
+            vb.set_auth_token(backend_data['auth'])
+            item_data = json.loads(cart_item['backend_data'])
+            res = vb.update_cart(item_data['line'], int(quantity))
+            net_price = Decimal(res['net_price'])
+            if backend_data['auth'] != res['auth']:
+                cart.update({'cart_data': json.dumps(internal_data)})
+        if net_price is not None:
+            stnd_price = Decimal(cart_item['price']) * Decimal(quantity)
+            print(f'Standard Price: {stnd_price} Net Price: {net_price}')
+            if net_price != stnd_price:
+                if net_price < stnd_price:
+                    raise Exception('Net price less than standard price for cart item: {}'.format(cart_item.sql_id()))
+                tax_diff = net_price - stnd_price
+                tax_diff = str(tax_diff)
+        #print('Cart Updates: {}'.format(updates))
+        return tax_diff
 
+    def backend_remove_cart_item(self, cart, backend_id, cart_item):
+        bkrec = sql_row('user_backend', id=backend_id)
+        if not bkrec.exists():
+            raise Exception('Invalid user backend: {}'.format(bkid))
+        urec = sql_row('user', id=bkrec['user_id'])
+        bkcfg = json.loads(bkrec['config_data'])
+        backend = bkrec['backend_name']
+        internal_data = json.loads(cart['cart_data'])
+        #pprint.pprint(internal_data)
+        if backend == 'vendure':
+            backend_data = internal_data['backend'][backend][backend_id]
+            vb = VendureBackend(None, URIRef(urec['merchant_uri']), bkcfg['vendure_url'])
+            vb.set_auth_token(backend_data['auth'])
+            item_data = json.loads(cart_item['backend_data'])
+            res = vb.remove_from_cart(item_data['line'])
+            if backend_data['auth'] != res['auth']:
+                backend_data['auth'] = res['auth']
+                updates = {'cart_data': json.dumps(internal_data)}
+                cart.update(updates)
  
     def add_cart_item(self, slug, quantity):
         qty = int(round(Decimal(quantity), 0))
@@ -200,9 +242,10 @@ class CatalogCart():
         item = sql_row('client_cart_item', cart_id=cart_id, entry_key=entry_key, product_index=index)
         if item.exists():
             newqty = item['quantity'] + qty
-            #self.backend_update_cart_item(cart, item['backend_id'], product, qty)
+            net_tax = self.backend_update_cart_item(cart, entry['backend_id'], item, newqty)
             item.update({
                 'quantity': newqty,
+                'net_tax': net_tax,
             })
         else:
             img = None
@@ -220,7 +263,7 @@ class CatalogCart():
                     img = product['image'][0]['url']
             offer = product['offers'][0]
             price = Decimal(offer['price'])
-            backend_data, net_tax = self.backend_add_cart_item(cart, entry['backend_id'], product, index, price, qty)
+            backend_data, net_tax = self.backend_add_cart_item(cart, str(entry['backend_id']), product, index, price, qty)
             sql_insert('client_cart_item', {
                 'cart_id': cart_id,
                 'backend_id': entry['backend_id'],
@@ -254,8 +297,13 @@ class CatalogCart():
         item = sql_row('client_cart_item', cart_id=cart_id, entry_key=entry_key, product_index=index)
         if not item.exists():
             raise Exception('Invalid cart entry key: {}'.format(slug))
+        entry = sql_row('entry', entry_key=item['entry_key'])
+        if not entry.exists():
+            raise Exception('Invalid entry')
+        net_tax = self.backend_update_cart_item(cart, str(entry['backend_id']), item, qty)
         item.update({
             'quantity': qty,
+            'net_tax': net_tax,
         })
         cart.update({
             'ts_updated': sql_now(),
@@ -272,6 +320,10 @@ class CatalogCart():
         cart_id = cart.sql_id()
         item = sql_row('client_cart_item', cart_id=cart_id, entry_key=entry_key, product_index=index)
         if item.exists():
+            entry = sql_row('entry', entry_key=item['entry_key'])
+            if not entry.exists():
+                raise Exception('Invalid entry')
+            self.backend_remove_cart_item(cart, str(entry['backend_id']), item)
             item.delete()
         cart.update({
             'ts_updated': sql_now(),
@@ -291,30 +343,27 @@ class CatalogCart():
         product = coder.decode_rdf(entry['external_uri'])
         return product
 
-    def prepare_checkout(self):
+    def prepare_checkout(self, spec):
         cart = self.build_cart()
         cart_id = cart.sql_id()
         internal_data = json.loads(cart['cart_data'])
+        payments = []
         if not internal_data.get('checkout_prepared', False):
             internal_data.setdefault('backend', {})
-            items = self.get_cart_items(cart_id, limit=1000, backend=True)
+            items = self.get_cart_items(cart_id, limit=1000)
             merchants = items['merchants']
             backends = {}
             for item in items['items']:
                 if item['backend_id'] not in backends:
                     backends[item['backend_id']] = {
-                        'products': [],
+                        'total': Decimal(0),
                         'merchant': merchants[item['merchant']],
                     }
-                product = self.get_entry_product(item['entry_key'])
-                backends[item['backend_id']]['products'].append({
-                    'product': product,
-                    'quantity': item['quantity'],
-                    'index': item['product_index'],
-                })
-                del item['product_index']
-                del item['backend_id']
-                del item['entry_key']
+                bkdata = backends[item['backend_id']]
+                item_total = Decimal(item['price']) * Decimal(item['quantity'])
+                if item['net_tax'] is not None:
+                    item_total = item_total + Decimal(item['net_tax'])
+                bkdata['total'] = bkdata['total'] + item_total
             for bkid in backends.keys():
                 backend_cart = backends[bkid]
                 bkrec = sql_row('user_backend', id=bkid)
@@ -322,22 +371,24 @@ class CatalogCart():
                     raise Exception('Invalid user backend: {}'.format(bkid))
                 bkcfg = json.loads(bkrec['config_data'])
                 backend = bkrec['backend_name']
-                internal_data['backend'].setdefault(bkid, {})
-                backend_data = internal_data['backend'][bkid]
+                backend_data = internal_data['backend'][backend][str(bkid)]
                 if backend == 'vendure':
                     vb = VendureBackend(None, URIRef(backend_cart['merchant']['id']), bkcfg['vendure_url'])
-                    backend_data.update(vb.prepare_checkout(backend_cart['merchant'], backend_cart['products']))
+                    vb.set_auth_token(backend_data['auth'])
+                    res = vb.prepare_checkout(backend_cart['merchant'], spec)
+                    print(f'Vendure Prepare Checkout: {res}')
+                payment_data = {}
+                payments.append({
+                    'method': 'atellixpay',
+                    'total': str(backend_cart['total']),
+                    'data': payment_data,
+                })
             internal_data['checkout_prepared'] = True
             cart.update({
                 'cart_data': json.dumps(internal_data),
                 'ts_updated': sql_now(),
             })
-            cart_items = items
-        else:
-            cart_items = self.get_cart_items(cart_id)
-        cart_updated = sql_row('client_cart', id=cart_id)
-        cart_data = cart_updated.data()
-        cart_data['cart_data'] = {}
-        cart_data['cart_items'] = cart_items
-        return cart_data
+        return {
+            'payments': payments
+        }
